requirements.txt
fastapi
uvicorn
PyJWT
python-dotenv
cryptography
sqlalchemy
psycopg2-binary
pydantic
email_validator
passlib[bcrypt]
bcrypt==4.0.1
alembic
emails
Jinja2
slowapi
